LayoutTests/sputnik/Conformance/12_Statement/12.9_The_return_Statement/S12.9_A1_T10-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Parse error
S12.9_A1_T10

PASS Expected parsing failure

TEST COMPLETE
